Convert all argument / return SFINAE into template argument SFINAE.
Similarly as done in Aug 2024 in Corrade. When these were a part of the
function signature, they ended up being encoded into the exported
symbol. There are still cases of StridedArrayView slice() having
enable_if in the signature, which amounts to about 18 kB symbols in all
libMagnum*-d.so libraries, but apart from that this is the state before:

    $ strings libMagnum*-d.so | grep enable_if | grep -v slice | wc -c
    29591

And this is after. All of those are coming from STL, thus from
old or deprecated APIs that still use std::vector, std::tuple and such,
and from the few std::sort() uses.

    $ strings libMagnum*-d.so | grep enable_if | grep -v slice | wc -c
    4103

In a non-deprecated build it's just this, which is a 10x reduction.
Can't really do much about these maybe exceút for implementing my own
swap() specializations (sigh?), but I think it's fine.

    $ strings libMagnum*-d.so | grep enable_if | grep -v slice | wc -c
    2904

I also made it consistently use

    typename std::enable_if<..., int>::type = 0

instead of

    class = typename std::enable_if<...>::type

because the former works correctly also in presence of overloads and
having it used consistently everywhere makes it easier to grep & change
later. All SFINAE is now also excluded from Doxygen output, because it
doesn't make much sense there. It's better to just explain the
restriction in words than with this nasty hack.