singles: add include guards for the implementation parts.
With seemingly innocent code like below, it happened that the
implementation part from MagnumMath.hpp got included twice because
MagnumMathBatch.hpp includes MagnumMath.hpp as a dependency, leading to
duplicate symbol definitions. This should be a completely valid use
case, so I'm adding guards to ensure the implementation is always only
added if not already.

    #define MAGNUM_MATH_IMPLEMENTATION
    #include <MagnumMath.hpp>
    #include <MagnumMathBatch.hpp>

A corresponding test case is going to get added to the singles repo.