doc: expand MSVC-/GCC-specific warning supporessions to clang-cl.
Strangely enough, that C4312 warning suppression now needs to be C4834
instead while I'm pretty sure it was correct back in March 2024 when I
added this in 4f7b57ffd6cb21dc34f982b741ff9b8655562b0f. C4312 is however
for "conversion of A to B of greater size", and in
4435877cf1cf91445c149c55aa22c94b9170788e I was fixing a lot of those, so
maybe I just used the same number by accident and because this
particular warning is extremely easy to miss, it was never correct in
the first place? Who knows.